+++ src/main/java/act/conf/AppConfigKey.java
@@ -29,8 +29,6 @@ import act.view.TemplatePathResolver;
 import act.view.View;
 import act.ws.DefaultSecureTicketCodec;
 import act.ws.SecureTicketCodec;
-import org.osgl.$;
-import org.osgl.exception.NotAppliedException;
 import org.osgl.util.E;
 import org.osgl.util.S;
 
@@ -1256,12 +1254,7 @@ public enum AppConfigKey implements ConfigKey {
     }
 
     public static void onApp(final App app) {
-        helper.classLoaderProvider(new $.F0<ClassLoader>() {
-            @Override
-            public ClassLoader apply() throws NotAppliedException, $.Break {
-                return app.classLoader();
-            }
-        });
+        helper.onApp();
     }
 
     /**