+++ src/main/java/act/util/Lazy.java
@@ -26,9 +26,12 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Mark an singleton class to be initialized after application started
+ * Mark an singleton class to be lazy initialized or a method
+ * that is not subject to eager load (e.g. the method in
+ * AppConfig)
+ *
  */
-@Target({ElementType.TYPE})
+@Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Lazy {
 }